DEFINE-COMMAND no longer defines a wrapper function.
This is because it makes little sense to have those for a command, which only really exists in the context of a parsed message and a client. It is suggested that the module writer write his module in such a way that it is still functionally accessible and have the commands be merely primitive wrappers for a user interface.

There is an alternate solution to this, which is to have the commands return functional values and define some kind of way to format the outputted data, allowing them to still be used in a functional context, but also work in a message context. However, this approach was rejected for the following reasons: first, doing so would require changing the internal definition to a query rather than an instruction and thus each call would involve quite a bit of overhead. Second, by having to choose a singular way of formatting the returned values we lose flexibility and would potentially have to fall back on errors to get desired output or perform other kludges to fix things up. Third, it would require changing the behaviour of queries to port over errors as well. While this part may still be done due to general usefulness it is currently not implemented.