Major change!
Consumers now act as their own event loops, and the default for handlers is to be added to the consumer, rather than the loops. Handlers now are also usually blocking, rather than queued. The old behaviour is still achievable on a per-handler basis, by changing the handler option :add-to-consumer to NIL.

This should significantly reduce the overhead on typical setups, as we now no longer need a thread per handler, but just per consumer.